Widget rerender optimization on setting change
